Remove inline queries for getting full post
Getting full posts did not work because long posts do not
fit in a single message so we have to send multiple messages
and implementing this logic introduces unnecessary complexity
because reading the full post in telegram is not convenient.
So we decided to just send an inline query to open post in browser.