stick with IHostedService
The problem with IHostedService is that the app does not start
until the StartAsync method starts. So if you have a long running method
the app won't start until StartAsync finishes. And that actually makes
IHostedService more suited to this scenario because the whole
app is a bot and nothing should started until the webhook is
ready to recieve requests