resolve initialization dependency issue
Addressed the chicken-and-egg problem of initializing the command and migrator
by manually creating the command without DI and introducing a factory method for
migrator creation. This approach decouples command initialization which does
not require DI from the migrator which requires it, allowing for the connection
string to be supplied as a parameter thus eliminating the need on having the
connection string before command creation.