conform to SRP principle
Having a single big MessageHandler class proved to be hard
to maintain and test, moving to CommandHandlers made it easier
to test, maintaint, and reason about the code.
The code migrated to using SimpleInjector due to MS.DI not supporting
registeration of open generics. Other changes were made because of
some incompatibility issues with SimpleInjector.